Add files via upload
Problem Occurs in Our institution's website:- Our Institution's website (Sityog Institute of Technology) is not a College friendly website, Because it doesn't uses CMS development which provides faculty members as well as principal to upload notices and other important details at a time without any kind of coding. Engineering students face the challenge of keeping track of multiple subjects and syllabus for each semester, which can be overwhelming and time- consuming. It is often difficult for them to find a centralized source for all the syllabus information, which can lead to confusion and inefficiency in their studies. This can also result in missed deadlines and poor academic performance, which can have long-term consequences for their careers. Moreover, the current process of accessing the syllabus is often manual, which can be prone to errors and inconsistencies, and can also cause delays in the dissemination of important information. Additionally, students may have difficulty understanding the syllabus content or may miss important updates or changes to the syllabus .To address these challenges, our Web app provides a comprehensive solution that provides engineering students with a centralized location for all syllabus information for each semester. By simplifying the process of accessing and understanding the syllabus, the app can help students stay organized, improve their academic performance, and reduce the risk of missed deadlines.