Problem Solved by my Project Smart Sityog
The primary objective of our project is to provide engineering students with a user-friendly and comprehensive Web app that contains the syllabus for all semesters. Specifically, our objectives are:
To develop a Web app that provides engineering students with a centralized
Location for all syllabus information for each semester, making it easier for them to access 
and keep track of their coursework.
* To ensure that the Web app is user-friendly and accessible to students of all technical 
backgrounds, including those who are not familiar with complex technical jargon.
* To create an app that is regularly updated with the latest syllabus information, including 
any changes or updates made by the university or college.
* To provide a platform for students to interact with their peers, share notes, and discuss 
coursework, which can improve their learning outcomes and academic performance.
* To provide a feature that allows students to set reminders for deadlines, exams, and 
assignments, which can help them stay organized and reduce the risk of missed deadlines.
* To provide a platform for faculty members or principal of the institution to upload the latest notices and also provide a platform for TPO(Training Placement Officer) to upload latest placement details at a time without doing any kind of coding.
Overall, our objectives aim to simplify the process of accessing and understanding the syllabus for engineering students, and to provide them with a tool that can help them stay organized and excel in their studies.